+++ extensions/cloudinary/actions/uploadFileFromDataPoint/index.ts
@@ -1,0 +1,1 @@
+export { uploadFileFromDataPoint } from './uploadFileFromDataPoint'